Update poke_redirect.py
Add some trailing data to some of the patches.  This doesn't change the output binary, but it makes the patches more easily reorderable, and the assembly a bit more symetric.

Note that the nop at the end of the SD functions isn't neccessary, it just makes reading the binary a bit simpler.